feat: Split writer into connection worker and wrapper, this is a (#1774)
prerequisite for multiplexing client
