+++ .github/.OwlBot.yaml
@@ -77,6 +77,7 @@ deep-preserve-regex:
 - "/google-cloud-bigquerystorage/src/main/java/com/google/cloud/bigquery/storage/v1.*/StreamWriterV2.java"
 - "/google-cloud-bigquerystorage/src/main/java/com/google/cloud/bigquery/storage/v1.*/Waiter.java"
 - "/google-cloud-bigquerystorage/src/main/java/com/google/cloud/bigquery/storage/v1/Exceptions.java"
+- "/google-cloud-bigquerystorage/src/main/java/com/google/cloud/bigquery/storage/v1/ConnectionWorker.java"
 
 deep-copy-regex:
 - source: "/google/cloud/bigquery/storage/(v.*)/.*-java/proto-google-.*/src"
